Revert "wip : normalized"
