+++ CMakeLists.txt
@@ -15,7 +15,6 @@ target_include_directories(goofycast PUBLIC
     "${PROJECT_SOURCE_DIR}/src/userinputs"
     "${PROJECT_SOURCE_DIR}/src/vec3"
     "${PROJECT_SOURCE_DIR}/src/raytrace"
-    "${PROJECT_SOURCE_DIR}/src/object"
 )
 
 add_subdirectory(./src/raymath)
@@ -24,6 +23,5 @@ add_subdirectory(./src/lodepng)
 add_subdirectory(./src/userinputs)
 add_subdirectory(./src/vec3)
 add_subdirectory(./src/raytrace)
-add_subdirectory(./src/object)
 
-target_link_libraries(goofycast PUBLIC raymath rayimage lodepng userinputs vec3 raytrace object hittable)
+target_link_libraries(goofycast PUBLIC raymath rayimage lodepng userinputs vec3 raytrace)
