+++ main.cpp
@@ -1,9 +1,42 @@
-#include "main.hpp"
+#include <iostream>
+#include <ctime>
+#include <sstream>
+#include <iomanip>
+#include <string>
+#include "Color.hpp"
+#include "Image.hpp"
+#include "Uinputs.hpp"
+#include "Vec3.hpp"
+#include "Raytrace.hpp"
 
-color ray_color(const raytrace& r, const hittable& world) {
-    hit_record rec;
-    if (world.hit(r, 0, infinity, rec)) {
-        return 0.5 * (rec.normal + color(1,1,1));
+using namespace std;
+
+std::string getCurrentTime() {
+    std::time_t t = std::time(nullptr);
+    std::tm tm = *std::localtime(&t);
+    std::ostringstream oss;
+    oss << std::put_time(&tm, "%Y%m%d_%H%M%S");
+    return oss.str();
+}
+
+double hit_sphere(const point3& center, double radius, const raytrace& r) {
+    vec3 oc = center - r.origin();
+    auto a = r.direction().length_squared();
+    auto b = dot(r.direction(), oc);
+    auto c = oc.length_squared() - radius*radius;
+    auto discriminant = b*b - a*c;
+    if (discriminant < 0) {
+        return -1.0;
+    } else {
+        return (-b - sqrt(discriminant)) / (2.0*a);
+    }
+}
+
+color ray_color(const raytrace& r) {
+    auto t = hit_sphere(point3(0,0,-1), 0.5, r);
+    if (t > 0.0) {
+        vec3 N = unit_vector(r.at(t) - vec3(0,0,-1));
+        return 0.5*color(N.x()+1, N.y()+1, N.z()+1);
     }
 
     vec3 unit_direction = unit_vector(r.direction());
@@ -25,6 +58,7 @@ int main() {
     image_height = (image_height < 1) ? 1 : image_height;
     Image image(image_width, image_height);
 
+
     auto focal_length = 1.0;
     auto viewport_height = 2.0;
     auto viewport_width = viewport_height * (double(image_width)/image_height);
@@ -38,16 +72,6 @@ int main() {
                              - vec3(0, 0, focal_length) - viewport_u/2 - viewport_v/2;
     auto pixel00_loc = viewport_upper_left + 0.5 * (pixel_delta_u + pixel_delta_v);
 
-    hittable_list scene;
-    scene.add(make_shared<sphere>(point3(0, 0, -1), 0.5));
-    scene.add(make_shared<sphere>(point3(0, -100.5, -1), 100));
-    scene.add(make_shared<sphere>(point3(1, 0, -1), 0.2));
-    scene.add(make_shared<sphere>(point3(-1, 0, -1), 0.2));
-    //scene.add(make_shared<sphere>(point3(0, 0, -.5), 0.5));
-    //scene.add(make_shared<sphere>(point3(0, 0, -1.5), 0.5));
-    //scene.add(make_shared<sphere>(point3(0.5, 0.2, -.5), 0.3));
-    //scene.add(make_shared<sphere>(point3(0, -100.5, -1), 100));
-
 
     std::cout << "P3\n" << image_width << " " << image_height << "\n255\n";
     for (int j = 0; j < image_height; j++) {
@@ -57,10 +81,12 @@ int main() {
             auto ray_direction = pixel_center - camera_center;
             raytrace r(camera_center, ray_direction);
 
-            color pixel_color = ray_color(r, scene);
+            color pixel_color = ray_color(r);
 
-            //std::clog << "i: " << i << ", j: " << j << ", image_width: " << image_width << ", image_height: " << image_height << "\n";
+            // Log des indices et des dimensions de l'image
+            std::clog << "i: " << i << ", j: " << j << ", image_width: " << image_width << ", image_height: " << image_height << "\n";
 
+            // Vérification des indices avant de définir le pixel
             if (i >= 0 && i < image_width && j >= 0 && j < image_height) {
                 image.SetPixel(i, j, pixel_color);
             } else {
