+++ src/vec3/Vec3.cpp
@@ -40,10 +40,6 @@ double vec3::length_squared() const {
     return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
 }
 
-vec3 vec3::normalized() const {
-    return *this / this->length();
-}
-
 std::ostream& operator<<(std::ostream& out, const vec3& v) {
     return out << v.e[0] << ' ' << v.e[1] << ' ' << v.e[2];
 }
